make python package configurable (#144)
Using amz linux image on ec2 the python packages do not follow the original hard coded package name 'python' but rather 'pythonXX' where XX is the version number. In order to use this package I needed the ability to define which version of python I have installed on my machine.