first work around the installer core
The installer should be renamed to instance (install sounds like an
action rather than a resource) and should be kept private as
implementation detail as the base class should handle everything.

Furthermore the installer should be kept private and all the validation
code around it should be dropped (most of it is duplicated to the
init.pp). And the package management should be moved into another
location.