Fix metapackage linking on Ubuntu 14.04
It's optional (and enabled by default) to manage dependencies of the
module needed during deployment with `package` declarations (see
`build_deps` and `::nodejs::pkgs`).

However the `rubygems` package which explicitly ensures that the `gem`
provider is functional on each host is a metapackage that will be
replaced by `ruby`. This "linking" doesn't work on Ubuntu 14.04 LTS
which is still supported (please refer to the forum for further
reference: https://askubuntu.com/questions/613074/package-rubygems-has-no-installation-candidate).

Now each OS with `$::osfamily == 'Debian'` uses `rubygems-integration`
to fix the broken linking.

Fixes #181