Drop hacky `ruby` setup
To evaluate the Puppet catalogue on the Puppet Master, the client
doesn't need `rubygems` as well. This was an ancient hack (maintained
over years) to avoid such a setup in test environments for Puppet such
as Vagrant.

Instead something like this might be implemented:

``` ruby
Vagrant.configure(2) do |config|
  # ...

  config.vm.provision :puppet do |puppet|
    puppet.module_path = ['puppet', ...]
  end
end
```

With the `manifests/default.pp`:

``` puppet
node default {
  contain ::nodejs

  package { ['ruby', 'rubygems']:
    ensure => present,
  } ->
  Class['::nodejs']
}
```