Refactor `install_dir` usage to avoid `::nodejs::params` references inside the module
The `::nodejs::params` class shall only be used as container with
sensitive defaults to provide a working setup on an average OS.

It should never be referenced directly inside the module's code, instead
`::nodejs` should pass it down to all internal APIs.

This is not a breaking change and much rather a cosmetic change as it
only affects private parts of the module and `::nodejs` has a parameter
with a default value.