Make the node.js source (by default nodejs.org/dist) configurable
As requested it might be helpful to specify your own location for either
source or binary distribution. Possible cases where this might be needed
are e.g. corporate environments without internet access or to deploy a
patched/forked version of NodeJS to Puppet nodes.

As the NodeJS distribution provides several ways to index the available
versions, the source_server behaves different when used with a custom
server.

```
class { '::nodejs':
  source => 'https://your-server.com/nodejs.tar.gz',
}
```

You can't specify a version, but you need to directly specify the source
ot be installed.

`make_install => true` can be used for source builds with Puppet as
well.

Fixes #185