Introduce acceptance testing of Puppet manifests with beaker
Due to several changes inside to the dependencies in the module this is
a rather drastic change. The module `puppetlabs-gcc` seems to be
abandoned (Puppet 3.x only support, last commit in 2015) which broke an
automated install with enabled dependency resolution (as done in
beaker). The logic of the module is now directly in
`::nodejs::instance::pkgs`.

To automate such testing processesin the future, `beaker` has been
introduced with `docker` as hypervisor (it's far more lightweight than a
full-blown `vagrant` setup).

Currently we only support Ubuntu 16.04 LTS in the testing matrix, but
whenever distros start to conflict with this module, the test suite can
be easily extended.

As `PUPPET_VERSION` is interpreted by beaker as well the test suite has
been split into two tasks (:test and :acceptance) as gem's version
constraint will break APT while provisioning docker.

Fixes #188